% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal.surface.R
\name{principal.surface}
\alias{principal.surface}
\title{Principal Surfaces}
\usage{
principal.surface(X, max.iter = 10, alpha = 0.6, N = 50)
}
\arguments{
\item{X}{data of size n x p}

\item{max.iter}{maximum number of iterations in the principal surface algorithm if the algorithm does not converge.}

\item{alpha}{span argument in loess() function}

\item{N}{creates an N^2 x p interpolated grid surface}
}
\description{
Principal Surfaces
}
